% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adissoma.R
\name{adissoma}
\alias{adissoma}
\title{Totais de colunas em uma tabela}
\usage{
adissoma(tabela, rotulo = "Total", funcao = "soma", dec = 0, exclui = NULL)
}
\arguments{
\item{tabela}{Objeto de estrutura matricial com no mínimo duas linhas e duas colunas}

\item{rotulo}{Rótulo para a linha da soma, o padrão é "Total".}

\item{funcao}{Operação a ser realizada, o padrão é a soma (função \code{\link{sum}}), outras opções são a média (invoca a função \code{\link{mean}}) com o argumento \code{funcao = "mean"} ou \code{funcao = "media"}) e a mediana, que invoca a função \code{\link{median}} com o argumento \code{funcao = "mean"} ou \code{funcao = "media"}. Os missings ("NAs") são ignorados (com \code{na.rm = TRUE} ao invocar a função).}

\item{dec}{Nº de decimais, padrão é 0.}

\item{exclui}{Linhas a excluir da operação (v. exemplo).}
}
\value{
A tabela (de classe \code{data.frame}) com a linha da soma ao final.
}
\description{
Em uma tabela, acrescenta uma linha com o total das colunas, ou a soma dos valores das colunas em linhas selecionadas.
}
\details{
É análoga a \code{\link{colSums}}, mas identifica as colunas numéricas na tabela e trabalha apenas sobre elas, sempre com \code{na.rm = TRUE}, isto é, desconsiderando os missings na soma.
}
\examples{
# Reproduzir uma tabela da mortalidade por "faixa etária detalhada" do DATASUS:
tabela <- "'Faixa Etária det'  Masc  Fem  Ign Total
                 '0 a 6 dias'   361   257   5   623
                '7 a 27 dias'   156   111   -   267
              '28 a 364 dias'   155   148   1   304
                 '1 a 4 anos'   144    88   -   232
                 '5 a 9 anos'    64    46   -   110
               '10 a 14 anos'    83    68   -   151
               '15 a 19 anos'   462   141   -   603
               '20 a 24 anos'   817   271   -  1088
               '25 a 29 anos'   880   393   -  1273
               '30 a 34 anos'  1155   561   -  1716
               '35 a 39 anos'  1508   870   1  2379
               '40 a 44 anos'  1954  1190   -  3144
               '45 a 49 anos'  2514  1681   -  4195
               '50 a 54 anos'  3488  2175   2  5665
               '55 a 59 anos'  5170  3265   2  8437
               '60 a 64 anos'  6557  4240   2 10799
               '65 a 69 anos'  7566  5370   1 12937
               '70 a 74 anos'  7989  6181   3 14173
               '75 a 79 anos'  7438  6508   4 13950
             '80 anos e mais' 14444 21159   7 35610"
tabela <- read.table(header = TRUE, na.strings = "-", text = tabela)
adissoma(tabela)
# Somar a faixa de 0-4 anos:
tabela2 <- rbind(adissoma(tabela[1:4,], rotulo = "0 a 4 anos")[5,],
                 tabela[-c(1:4),])
adissoma(tabela2) 

# Pode-se ter o mesmo resultado com 'colSums()', mas com mais trabalho: 
cbind("Faixa etária" = c(tabela2[[1]], "Total"), 
      rbind(tabela2[-1], colSums(tabela2[-1], na.rm = TRUE)))

x <- xtabs(~ fxetar.det + sexo, obitosRS2019)
x
cbind(rownames(x),
      x |>
        unclass() |> 
        data.frame()  ) |> 
  adissoma() |>
  adissoma(funcao = "media", exclui = 34) |> # Exclui a linha 34 com o total
  adissoma(funcao = "mediana", exclui = 34:35)

class(x)

}
\seealso{
\code{\link{colSums}}
}
