% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolero.R
\name{bolero}
\alias{bolero}
\title{Bolero: tabelas 2x2}
\usage{
bolero(independente, dependente = NULL, dec = 2, dnn = NULL)
}
\arguments{
\item{independente}{Variável independente}

\item{dependente}{Variável dependente}

\item{dec}{No. de decimais}

\item{dnn}{Nome das variáveis a ser impresso no 'output'}
}
\value{
Um objeto da classe \code{list} com as tabelas de frequências absolutas e relativas, razão de probabilidades e de odds, com seus intervalos de confiança e valores-p.
}
\description{
Analisa uma tabela 2x2 e apresenta um output com rótulos em português
}
\examples{
with(RDRS2019, bolero(SEXO, MORTE))
with(RDRS2019, bolero(SEXO, MORTE, dnn = c("Sexo", "Óbito")))

}
