% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestados.R
\name{codestados}
\alias{codestados}
\title{Codigos, siglas e nomes dos estados brasileiros}
\usage{
codestados()
}
\description{
Lista as siglas e nomes dos estados e Distrito Federal do Brasil segundo seu código no IBGE. Com isso, permite a decodificação do código do IBGE para a sigla ou nome da Unidade da Federação.
}
\examples{
codestados()
codestados()[,2]
codestados()[,3]

}
\author{
Fúlvio B. Nedel
}
