% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descreve.R
\name{descreve}
\alias{descreve}
\title{Describes a numerical variable}
\usage{
descreve(
  x,
  dec = 2,
  na.rm = TRUE,
  data = NULL,
  histograma = TRUE,
  breaks = "Sturges",
  freq = TRUE,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  linhas = 2,
  curva = TRUE,
  densidade = FALSE,
  col.dens = 1,
  col = "yellow2",
  col.curva = "DarkGreen",
  col.media = 2,
  col.dp = col.media,
  col.mediana = 4,
  legenda = TRUE,
  lugar = "topright",
  lty.curva = 2,
  lwd.curva = 1,
  lty.dens = 3,
  lwd.dens = 2,
  lty = NULL,
  lwd = NULL,
  cex = NULL,
  ...
)
}
\arguments{
\item{x}{A numeric vector}

\item{dec}{The number of digits}

\item{na.rm}{TRUE (default) removes missings}

\item{data}{Optional argument. Data frame containing \code{x}. Defaults to NULL.}

\item{histograma}{TRUE (default) draws a histogram}

\item{breaks}{number of breaks, according to \code{\link{hist}}; the default method is 'Sturges'}

\item{freq}{TRUE (default) set y axis to frequency; FAlSE set it to density}

\item{main}{Graphic title}

\item{xlab}{Label for x axis}

\item{ylab}{Label for y axis}

\item{linhas}{TRUE (default) draws vertical lines with median, mean, and SD +1 position at the histogram}

\item{curva}{TRUE (default) draws the expected normal curve}

\item{densidade}{defaults to FALSE; turn to TRUE to draw the density line}

\item{col.dens}{colour of density line; defaults to 1 (black)}

\item{col}{colour of histogram bars; defaults to 'yellow2'}

\item{col.curva}{colour of normal curve; defaults to 'DarkGreen'}

\item{col.media}{colour of mean line; defaults to 2 (red)}

\item{col.dp}{colour of SD lines; defaults to col.media}

\item{col.mediana}{colour of median line; defaults to 4 (blue)}

\item{legenda}{TRUE (default) draws a legend}

\item{lugar}{legend position; defaults to 'toprigt'}

\item{lty.curva}{line type of normal curve; defaults to 2}

\item{lwd.curva}{line width of normal curve; defautls to 1}

\item{lty.dens}{line type of density curve; defaults to 3}

\item{lwd.dens}{line width of density curve; defaults to 2}

\item{lty}{line type for the histogram; defaults to NULL}

\item{lwd}{line width for the histogram; defaults to NULL}

\item{cex}{font size}

\item{...}{takes paramters from other functions}
}
\description{
Performs a 'complete' description of a numerical variable, continuous or integer, and plots a histogram maybe with lines for mean, median, 1 and 2 sd from and to mean and normal probability curve for the plotted parameters.
}
\seealso{
\code{\link{hist}} and \code{\link{par}} for graphic paramters
}
