% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{ed}
\alias{ed}
\title{Seleciona as últimas colunas de um banco de dados}
\usage{
ed(x, n = 5)
}
\arguments{
\item{x}{Banco de dados (objeto da classe \code{data.frame}, \code{matrix} ou \code{table})}

\item{n}{Número de colunas a selecionar (por padrão são 5)}
}
\description{
Enquanto \code{\link{tail}} mostra as últimas linhas de uma matriz ou "data frame", \code{ed()} mostra as últimas colunas de uma matriz ou "data frame".
}
\examples{
ed(RDRS2019) |> # As últimas 5 colunas do banco
  head()
ed(RDRS2019, 8) |> # As últimas 8 colunas do banco
  head() 

}
