% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enutri.R
\name{enutri}
\alias{enutri}
\title{Classificação do estado nutricional de adultos}
\usage{
enutri(data = NULL, peso, altura, imc = NULL)
}
\arguments{
\item{data}{Banco de dados com a(s) variável(is)}

\item{peso}{Peso em Kg}

\item{altura}{Altura em m}

\item{imc}{Índice de Massa Corporal}
}
\value{
Um fator com as categorias "Baixo peso", "Eutrófico", "Sobrepeso" ou "Obesidade".
}
\description{
Classifica o estado nutricional da pessoa segundo valores do Índice de Massa Corporal (IMC), usando os pontos de corte propostos pela Organização Mundial da Saúde. Pode ser informado um vetor com valores do IMC ou um vetor com o peso (Kg) e outro com altura (m^2).
}
\examples{
data("hasdm") 
enutri(hasdm) |> 
  tabuleiro()

}
