% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folhinha.R
\name{folhinha}
\alias{folhinha}
\title{Desenha um calendário}
\usage{
folhinha(
  inicio,
  fim = NULL,
  semanas = 18,
  diasem,
  feriados = NULL,
  lang = "pt",
  size = 3,
  coraula = "steelblue",
  corfer = "antiquewhite1",
  corfimde = "antiquewhite2",
  corNA = "white"
)
}
\arguments{
\item{inicio}{Data em formato YYYYMMDD (em caractere ou numeral) com o primeiro dia de aulas.}

\item{fim}{Data em formato YYYYMMDD (em caractere ou numeral) com o último dia do calendário.}

\item{semanas}{Número de semanas letivas. Por padrão são 18.}

\item{diasem}{Dias da semana das aulas, abreviados e em minúsculas.}

\item{feriados}{Vetor com uma lista de datas para marcar feriados e dias não letivos.}

\item{lang}{Idioma para os dias da semana, "pt" (padrão) para português ou "es" para castelhano.}

\item{size}{Tamanho da letra do dia no gráfico.}

\item{coraula}{Cor de fundo para os dias de aula.}

\item{corfer}{Cor de fundo para os feriados e dias não letivos.}

\item{corfimde}{Cor de fundo para os fins-de-semana.}

\item{corNA}{Cor de fundo para o período não letivo.}
}
\value{
Objeto ggplot2 com o desenho do calendário.
}
\description{
Útil para visualizar cronogramas de aulas
}
\examples{
folhinha(inicio = 20250305, diasem = c("seg", "qua", "sex"))
folhinha(inicio = 20250311, fim = 20250725, diasem = c("ter", "qui"))
folhinha(inicio = 20250304, sem = 15,
   feriados = c("20250418", "2025/05/01", "2025-06-19"),
   diasem = c("mar", "jue"), lang = "es",
   size = 2.5, coraula = "yellow", corfer = "tomato3", corfimde = "white")

}
