% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatL.R
\name{formatL}
\alias{formatL}
\title{Números em formato latino}
\usage{
formatL(x, digits = 1, format = "f", ...)
}
\arguments{
\item{x}{Um número ou vetor com números}

\item{digits}{Número de decimais}

\item{format}{Formato numérico (ver \code{\link{formatC}})}

\item{...}{Permite outros argumentos da função \code{\link{formatC}}}
}
\description{
Transforma números em caráter com ponto (.) como separador de milhar e vírgula (,) como separador decimal, alterando os padrões da função \code{\link{formatC}} para \code{big.mark = "."} e \code{decimal.mark = ","}.
}
\examples{
formatL(1234.5678)
formatL(rnorm(5), digits = 2)
}
\seealso{
\code{\link{formatC}}, \code{\link{format}}
}
