% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxetar_quinq.R
\name{fxetar_quinq}
\alias{fxetar_quinq}
\title{Age groups in five years, for DATASUS data}
\usage{
fxetar_quinq(x)
}
\arguments{
\item{x}{Age group with 34 levels: yearly from 0 to 19 years, five year from 20 to 80, and 80 and more (classificação "faixa etária detalhada" do DATASUS)}
}
\description{
Recode the "detailed age groups" variables of DATASUS tables into a variable with five years age groups, from 0-4 to 80 and more (17 levels)
}
