% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pir.R
\name{ggplot_pir}
\alias{ggplot_pir}
\title{Pirâmides populacionais}
\usage{
ggplot_pir(
  banco,
  idade,
  sexo,
  populacao = NULL,
  catsexo = c("masc", "fem"),
  cores = c("steelblue", "hotpink"),
  nsize = 3.5
)
}
\arguments{
\item{banco}{Um \code{data frame} com população ou casos por sexo e faixa etária.}

\item{idade}{Nome da variável com a idade ou faixa etária deve ir entre aspas.}

\item{sexo}{Nome da variável com o sexo; deve ir entre aspas e ser um \code{factor}.}

\item{populacao}{Nome da variável (entre aspas) com a população ou casos, se houver; por padrão é \code{NULL}, isto é, a função espera um banco de registros individuais e calcula a frequência por sexo e faixa etária indicada.}

\item{catsexo}{Vetor com o nome das categorias da variável sexo; padrão é \code{c("masc", "fem")}; o sexo masculino deve ser a primeira categoria.}

\item{cores}{Cores das barras, para as categorias masculino e feminino. O padrão é \code{c("darkblue", "violetred")}.}

\item{nsize}{Tamanho do texto com o nº total de habitantes. O padrão é 3.5.}
}
\description{
Pirâmides populacionais
}
\examples{
data("POPBR12")
str(POPBR12)
# "SEXO" é da classe 'integer'; 
# além disso, os primeiros 19 anos de idade estão em faixas anuais 
# e os demais (até 80) em faixas quinquenais.
\dontrun{
ggplot_pir(POPBR12, "FXETARIA", "SEXO", "POPULACAO")
}
# não dá o gráfico desejado.

# Transformando o sexo em fator e reclassificando a faixa etária:
require(dplyr)
POPBR12 <- POPBR12 \%>\% 
  mutate(SEXO = factor(SEXO, labels = c("masc", "fem")),
         FXETAR5 = fxetar.det_pra_fxetar5(FXETARIA)) 
ggplot_pir(POPBR12, "FXETAR5", "SEXO", "POPULACAO")

# Cerro Largo, RS
ggplot_pir(POPBR12[POPBR12$MUNIC_RES==430520, ], 
           "FXETAR5", "SEXO", "POPULACAO")


# ## Mortalidade por anos completos de vida

# Se o banco é de registros individuais, sem uma variável com a 
# contagem da população, a função conta as frequências em cada 
# sexo e faixa etária específicos. 

data("obitosRS2019")
str(obitosRS2019)

obitosRS2019 \%>\%
 mutate(sexo = factor(sexo, levels = c("masc", "fem"))) \%>\% 
 ggplot_pir("idade", "sexo")
 
# O comando acima devolve o mesmo gráfico que o comando abaixo: 

obitosRS2019 \%>\%
 mutate(sexo = factor(sexo, levels = c("masc", "fem"))) \%>\%
 count(idade, sexo) \%>\% # linha desnecessária
 ggplot_pir("idade", "sexo", "n")

}
