% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{hasdm}
\alias{hasdm}
\title{Usuários da APS com hipertensão ou diabete}
\format{
An object of class \code{data.frame} with 384 rows and 22 columns.
}
\usage{
hasdm
}
\description{
Usuários da APS com hipertensão ou diabete
}
\keyword{datasets}
