% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_popbr.R
\name{ler_popbr}
\alias{ler_popbr}
\title{Ler arquivos POPBR????.DBF}
\usage{
ler_popbr(pop, ...)
}
\arguments{
\item{pop}{População. Nome do arquivo a ser lido.}

\item{...}{outros argumentos da função `foreign`}
}
\description{
Lê os arquivos de população do DATASUS e cria uma variável com a faixa etária quinquenal
}
