% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamostra.R
\docType{data}
\name{mensagem.alfa}
\alias{mensagem.alfa}
\title{mensagem.alfa}
\format{
An object of class \code{character} of length 1.
}
\usage{
mensagem.alfa
}
\description{
mensagem.alfa
}
\keyword{datasets}
