% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamostra.R
\docType{data}
\name{mensagem.erro}
\alias{mensagem.erro}
\title{mensagem.erro}
\format{
An object of class \code{character} of length 1.
}
\usage{
mensagem.erro
}
\description{
mensagem.erro
}
\keyword{datasets}
