% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda}
\alias{moda}
\title{Computes the mode of a variable}
\usage{
moda(x, tipo = "int1", count = FALSE)
}
\arguments{
\item{x}{A vector. Maybe numeric, character or factor}

\item{tipo}{Defines the method. `int1` (the default) find the most frequent value in an unimodal distribution; `int2` finds the two most frequent values in a bimodal distribution; `dens` computes the value of most density in a numeric variable. If the variable is continuous, with only unique values, `tipo` will be coerced to `dens`}

\item{count}{Should the frequencies of values be described?}
}
\description{
Calcula a moda de uma variável
}
