% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numescrito.R
\name{numescrito}
\alias{numescrito}
\title{Escreve números em texto, em português}
\usage{
numescrito(x)
}
\arguments{
\item{x}{Número a ser convertido}
}
\description{
Dado um número entre 0 e 12, devolve seu valor em texto em português.
}
\examples{
numescrito(0)
numescrito(3)
numescrito(-2)
numescrito(20)
paste(numescrito(2), "+", numescrito(3), "=", numescrito(2+3))

}
