% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{obitosRS2019}
\alias{obitosRS2019}
\title{Registros de óbito}
\format{
Banco de dados com amostra aleatória de 10.000 DO de 2019 de residentes do RS, registrados no Sistema de Informação de Mortalidade (SIM). O banco foi previamente trabalhado para a decodificação da idade, de modo que a idade represente a idade em anos completos (0 para < 1 ano), e para a definição de missings e rótulos no campo sexo. Essas variáveis foram acrescentadas (com nome em letra minúscula, \code{idade} e \code{sexo}), sendo mantidas as originais (\code{IDADE}, \code{SEXO}).
}
\source{
Sistema de Informações sobre Mortalidade (SIM)
}
\usage{
obitosRS2019
}
\description{
Amostra de Declarações de Óbito (DO) de residentes no RS, 2019.
}
\keyword{datasets}
