% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{PASC2512}
\alias{PASC2512}
\title{Procedimentos Ambulatoriais do SUS}
\format{
Banco de dados de exemplo de um arquivo "PA" do SIASUS. Contém uma amostra aleatória de 1.000 registros do arquivo "PASC2512.dbc" com os procedimentos do Acre no mês de competência 12/2025.
}
\usage{
PASC2512
}
\description{
Procedimentos Ambulatoriais do SUS
}
\keyword{datasets}
