% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.histobox.R
\name{plot.histobox}
\alias{plot.histobox}
\title{Método para histobox}
\usage{
\method{plot}{histobox}(x, ...)
}
\arguments{
\item{x}{Uma variável numérica}

\item{...}{Permite a inclusão de argumentos para \code{\link{hist}}}
}
\description{
Método para histobox
}
