% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{poprs}
\alias{poprs}
\title{População por sexo e faixa etária. Municípios gaúchos, 1980, 2010, 2020.}
\format{
Banco de dados com estimativas populacionais por sexo e faixa etária para os municípios do Rio Grande do Sul em 1980, 2010 e 2020. Criado com as funções \code{\link[csapAIH]{ler_popbr}} \code{\link[csapAIH]{popbr2000_2021}}.
}
\usage{
poprs
}
\description{
População por sexo e faixa etária. Municípios gaúchos, 1980, 2010, 2020.
}
\keyword{datasets}
