% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{POPRS2019}
\alias{POPRS2019}
\title{População por sexo e faixa etária. Municípios gaúchos, 2019.}
\format{
Banco de dados com estimativas populacionais por sexo e faixa etária para os municípios do Rio Grande do Sul em 2019. Criado com a função \code{\link[csapAIH]{popbr2000_2021}}, que por sua vez usa os dados tabulados do DATASUS e importados pelo pacote \code{brpop}, de Raphael Saldanha.
}
\usage{
POPRS2019
}
\description{
População por sexo e faixa etária. Municípios gaúchos, 2019.
}
\keyword{datasets}
