% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.descreve.R
\name{print.descreve}
\alias{print.descreve}
\title{Imprime o resultado da funcao \code{\link{descreve}}}
\usage{
\method{print}{descreve}(x, print = "output", ...)
}
\arguments{
\item{x}{Um objeto da classe \code{descreve}, output de \code{\link{descreve}}}

\item{print}{Modo de apresentação; \code{print = "tabela"} retorna uma tabela com as estatísticas}

\item{...}{Não sei se serve de algo mas parece que precisa}
}
\description{
Imprime o resultado da funcao \code{\link{descreve}}
}
\concept{descreve}
