\name{Rcoisas-package}
\alias{Rcoisas-package}
\alias{Rcoisas}
\docType{package}
\title{
\packageTitle{Rcoisas}
}
\description{
\packageDescription{Rcoisas}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rcoisas}
\packageIndices{Rcoisas}
Funções minhas, pros arquivos do DATASUS, pra legendas em português, etc.
}
\author{
\packageAuthor{Rcoisas}

Maintainer: \packageMaintainer{Rcoisas}
}
%\references{
%% ~~ Literature or other references for background information ~~
%}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%% ~~ documentation directory ~~
\keyword{ package }
%\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%% ~~ simple examples of the most important functions ~~
%}
