% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{RDRS2019}
\alias{RDRS2019}
\title{Arquivo da AIH - RS, 2019}
\format{
Banco de dados com amostra aleatória de 10.000 internações dos "arquivos reduzidos" (RD*.DBC) das Autorizações de Internação Hospitalar (AIH) pagas pelo SUS em hospitais do RS em 2019 ("ano de competência" = 2019). O banco é formado pela junção (adição de linhas) de cada um dos arquivos RDRS1901.DBC, ..., RDRS1912.DBC.
}
\source{
Bases de Dados do Sistema de Informações Hospitalares do SUS (BD-SIH/SUS)
}
\usage{
RDRS2019
}
\description{
Internações hospitalares pagas pelo SUS no RS em 2019.
}
\keyword{datasets}
