% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigtap.R
\name{sigtap}
\alias{sigtap}
\title{Classifica os procedimentos do SISAB em grandes grupos de procedimentos da tabela SIGTAP.}
\usage{
sigtap(df, varproc = "CO_PROCEDIMENTO", nomeproc = FALSE)
}
\arguments{
\item{df}{Banco de dados com os procedimentos. Parâmetro obrigatório não definido \emph{a priori}.}

\item{varproc}{Nome da variável com o código do procedimento. Por padrão é "CO_PROCEDIMENTO" e deve vir entre aspas, simples ou duplas.}

\item{nomeproc}{Deve-se criar uma variável com o nome do procedimento?}
}
\description{
Classifica os procedimentos do SISAB em grandes grupos de procedimentos da tabela SIGTAP.
}
\note{
Agradeço a Matheus Pacheco Andrade, da Gerência de Informação da Secretaria Municipal de Saúde de Florianópolis, SC, por ter encontrado as tabelas no servidor FTP do DATASUS.
}
\examples{
\dontrun{
  sigtap(df = procedimentos, 'CO_PROCEDIMENTO') |> head()
  sigtap(procedimentos) |> head()
  sigtap(procedimentos, nomeproc = TRUE) |> head()
}
data("PASC2512")
sigtap(PASC2512, "PA_PROC_ID") |> head()
sigtap(PASC2512, "PA_PROC_ID", nomeproc = TRUE) |> head()

# A lista completa de códigos classificados em grupos pode ser conseguida com:
sigtap(Rcoisas:::tabproc) |> head()

}
