% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.tabela.R
\name{spss.tabela}
\alias{spss.tabela}
\title{Tabela de frequências semelhante ao SPSS}
\usage{
spss.tabela(varcat, digits = 1)
}
\arguments{
\item{varcat}{Uma variável categórica}

\item{digits}{No. de digits *(em \code{\link{round}}, mudar pra formatC*)}
}
\description{
Tabela de frequências semelhante ao SPSS
}
