% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuleiro.R
\name{tabuleiro}
\alias{tabuleiro}
\title{Tabela de frequências univariada}
\usage{
tabuleiro(x, digits = 1, total = TRUE, cum = TRUE, data = NULL, ...)
}
\arguments{
\item{x}{O vetor a ser tabulado.}

\item{digits}{nº de decimais na tabela. O padrão é 1.}

\item{total}{TRUE (default) apresenta o total de categorias na tabela.}

\item{cum}{TRUE (default) apresenta as frequências acumuladas das cateogrias.}

\item{data}{Argumento opcional. Banco de dados contendo \code{x}. O padrão é NULL.}

\item{...}{permite o uso de argumentos da função \code{\link{table}}.}
}
\description{
Constrói uma tabela com distribuição de frequências brutas, relativas e acumuladas, com rótulos em português.
}
\examples{
set.seed(1)
x <- rbinom(100000, 3, .25)

tabuleiro(x)

# Sem o total
tabuleiro(x, total = FALSE)

# Sem as frequências acumuladas
tabuleiro(x, cum = FALSE)

# Oculta a frequência acumulada absoluta e mantém a \% acumulada
tabuleiro(x, total = FALSE)[,-3]

}
