% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuleiro2.R
\name{tabuleiro2}
\alias{tabuleiro2}
\title{Tabela de frequências semelhante ao output do SPSS}
\usage{
tabuleiro2(varcat, digits = 1)
}
\arguments{
\item{varcat}{Uma variável categórica}

\item{digits}{No. de decimais}
}
\description{
Tabela de frequências semelhante ao output do SPSS
}
\examples{
data("iris") 
varcat <- iris$Species
varcat[1:7] <- NA # criar missings
tabuleiro2(varcat)

}
