% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamostra.R
\name{tamostra}
\alias{tamostra}
\alias{tamostra.zmedia}
\alias{tamostra.tmedia}
\alias{tamostra.prop}
\alias{tamostra.2prop}
\title{Funcoes para calcular o tamanho de amostra}
\usage{
tamostra.zmedia(dp, erro, alfa = 0.05, perdas = 0)

tamostra.tmedia(dp, erro, pop = 10000, alfa = 0.05, perdas = 0)

tamostra.prop(prop, erro, alfa = 0.05, perdas = 0)

tamostra.2prop(p1, p2, alfa = 0.05, beta = 0.2, perdas = 0)
}
\arguments{
\item{dp}{desvio-padrão; argumento apenas nas estimações de média}

\item{erro}{erro aceitável, em valor absoluto nas estimações de média e em proporção (0 -- 1) ou em pontos percentuais (1 -- 100) nas estimações de proporção de uma medida na população; veja \code{detalhes}}

\item{alfa}{nível de significância, (1 - confiança); 0 < alfa < 1 (deve ser maior que zero e menor que um); default = 0,05}

\item{perdas}{fator de correção para perdas, em proporção (default = 0)}

\item{pop}{tamanho da população (default = 10000)}

\item{prop}{proporção esperada}

\item{p1}{proporção na população 1}

\item{p2}{proporção na população 2}

\item{beta}{probabilidade de erro tipo II (1-poder)}
}
\description{
Calcula o tamanho de amostra para estimar médias e proporções e para comparar duas proporções

tamostra.zmedia Estimar médias com variância conhecida (teste Z)

tamostra.tmedia Estimar médias com variância desconhecida (teste t)

tamostra.prop Estimar proporções

Comparar duas proporções
}
\details{
Veja o caso de cada função específica

Quando o valor de \code{erro} é menor que um, é entendido como proporção; quando maior ou igual a um, é entendido como pontos percentuais; \code{erro >= 100} retorna erro

### Precisa esta linha? ###

### Precisa esta linha? ###

### Precisa esta linha? ###
}
\section{Functions}{
\itemize{
\item \code{tamostra.zmedia}: Estimar médias com variância conhecida (teste Z)

\item \code{tamostra.tmedia}: Estimar médias com variância desconhecida (teste t)

\item \code{tamostra.prop}: Estimar proporções

\item \code{tamostra.2prop}: Comparar duas proporções
}}

