% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tango.R
\name{tango}
\alias{tango}
\title{Apresenta uma tabela multidimensional}
\usage{
tango(
  independente,
  dependente,
  dec = 2,
  nomes = c(substitute(independente), substitute(dependente)),
  p.value = trend
)
}
\arguments{
\item{independente}{Variavel independente}

\item{dependente}{Variavel dependente}

\item{dec}{no. de decimais}

\item{nomes}{nomes para as filas e colunas}

\item{p.value}{?????????????}
}
\description{
Apresenta uma tabela n x 2
}
