% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolero.R
\name{bolero}
\alias{bolero}
\title{Bolero: tabelas 2x2}
\usage{
bolero(independente, dependente = NULL, dec = 2, dnn = NULL)
}
\arguments{
\item{independente}{Variável independente}

\item{dependente}{Variável dependente}

\item{dec}{No. de decimais}

\item{dnn}{Nome das variáveis}
}
\description{
Analisa uma tabela 2x2 e apresenta um output com rótulos em português
}
