% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descreve.R
\name{descreve}
\alias{descreve}
\title{Descreve uma variável numérica}
\usage{
descreve(
  x,
  by = NULL,
  dec = 2,
  na.rm = TRUE,
  data = NULL,
  histograma = TRUE,
  breaks = "Sturges",
  freq = TRUE,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  linhas = 2,
  curva = TRUE,
  densidade = FALSE,
  col.dens = 1,
  col = "yellow2",
  col.curva = "DarkGreen",
  col.media = 2,
  col.dp = col.media,
  col.mediana = 4,
  legenda = TRUE,
  lugar = "topright",
  lty.curva = 2,
  lwd.curva = 1,
  lty.dens = 3,
  lwd.dens = 2,
  lty = NULL,
  lwd = NULL,
  cex = NULL,
  print = "output",
  ...
)
}
\arguments{
\item{x}{Um vetor numérico}

\item{by}{tentar incluir esse argumento}

\item{dec}{Número de dígitos}

\item{na.rm}{TRUE (default) remove os missings}

\item{data}{Argumento opcional. Banco de dados contendo \code{x}. O padrão é NULL.}

\item{histograma}{TRUE (default) desenha um histograma}

\item{breaks}{Número de divisões, de acordo com \code{\link{hist}}; o método padrão é Sturges'}

\item{freq}{TRUE (default) define o eixo y como frequência; FAlSE o define como densidade}

\item{main}{Título do gráfico}

\item{xlab}{Rótulo do eixo x}

\item{ylab}{Rótulo do eixo y}

\item{linhas}{TRUE (default) desenha linhas verticais com a média, mediana e média + 1 ou 2 DP}

\item{curva}{TRUE (default) desenha a curva normal esperada}

\item{densidade}{O padrão é FALSE; passe para TRUE para desenhar a curva de densidade da distribuição}

\item{col.dens}{Cor da curva de densidade; o padrão é 'black'}

\item{col}{Cor das barras do histograma; o padrão é 'yellow2'}

\item{col.curva}{Cor da curva normal; o padrão é 'DarkGreen'}

\item{col.media}{Cor da linha da média; o padrão é 'red'}

\item{col.dp}{Cor das linhas do SP; por padrão é a mesma que col.media ('red')}

\item{col.mediana}{Cor da linha da mediana; o padrão é 'blue'}

\item{legenda}{TRUE (default) desenha a legenda}

\item{lugar}{Posição da legenda; padrão é 'topright'}

\item{lty.curva}{Tipo da linha da curva normal; padrão é 2}

\item{lwd.curva}{Largura da linha da curva normal; padrão é 1}

\item{lty.dens}{Tipo da linha da curva de densidade; padrão é 3}

\item{lwd.dens}{Largura da linha da curva de densidade; padrão é 2}

\item{lty}{Tipo da linha para o histograma; padrão é NULL}

\item{lwd}{Largura da linha para o histograma; padrão é NULL}

\item{cex}{Tamanho da fonte}

\item{print}{Modo de apresentação; \code{print = "tabela"} retorna uma tabela com as estatísticas}

\item{...}{Toma parâmetros de outras funções utilizadas}
}
\description{
Executa a descrição "completa" de uma variável numérica, contínua ou discreta, e desenha um histograma, possivelmente com linhas para a média, mediana e distância da média até 1 ou 2 desvios-padrão, além da curva de probabilidade Normal para os parâmetros apresentados.
}
\examples{
descreve(rnorm(100))
descreve(rnorm(100), breaks = 'Scott')
descreve(rnorm(100), breaks = 'Scott', linhas = FALSE, densidade = TRUE, col = 'skyblue') 
descreve(rnorm(100), print = "tabela")
obitosRS2019$sexo <- factor(obitosRS2019$sexo)
descreve(obitosRS2019$idade)
\dontrun{
# NÃO FUNCIONA!
descreve(idade, obitosRS2019$sexo, data = obitosRS2019)
# Esse funciona
with(obitosRS2019, descreve(idade, sexo))
}

 
}
\seealso{
\code{\link{hist}} and \code{\link{par}} para os parâmetros gráficos
}
