% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxetar.det_pra_fxetar5.R
\name{fxetar.det_pra_fxetar5}
\alias{fxetar.det_pra_fxetar5}
\title{Transforma a "faixa etária detalhada" (DATASUS) em 17 faixas quinquenais.}
\usage{
fxetar.det_pra_fxetar5(x)
}
\arguments{
\item{x}{Um vetor com a idade categorizada nas seguintes faixas etárias:
\itemize{
  \item anos completos até 19 anos; 
  \item faixas quinquenais até 75-79 anos; e 
  \item 80 e + anos.
}}
}
\value{
Um vetor com a idade categorizada em 17 faixas etárias: quinquenais de 0 a 79 anos e 80 e + anos de idade.
}
\description{
Reclassifica as idades < 20 anos em faixas etárias quinquenais.
}
\details{
Falar das faixas etárias do tabnet e tabwin
}
\examples{
data("POPBR12")
str(POPBR12)
POPBR12$FXETAR5 <- fxetar.det_pra_fxetar5(POPBR12$FXETARIA)
str(POPBR12)
levels(POPBR12$FXETAR5)

}
\keyword{DATASUS}
