% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pir.R
\name{ggplot_pir}
\alias{ggplot_pir}
\title{Pirâmides populacionais com os arquivos de população
disponibilizados pelo DATASUS}
\usage{
ggplot_pir(
  banco,
  idade,
  sexo,
  populacao = NULL,
  catsexo = c("masc", "fem"),
  cores = c("darkblue", "violetred"),
  nsize = 3.5
)
}
\arguments{
\item{banco}{Um \code{data frame} com população ou casos por sexo e faixa etária.}

\item{idade}{nome da variável com a idade ou faixa etária.}

\item{sexo}{nome da variável com o sexo; deve ser um \code{factor}.}

\item{populacao}{nome da variável com a população ou casos, se houver; por padrão é \code{NULL} e a função calcula a frequência por sexo e faixa etária indicada.}

\item{catsexo}{vetor com o nome das categorias da variável sexo; padrão é \code{c("masc", "fem")}; o sexo masculino deve ser a primeira categoria.}

\item{cores}{Cores das barras, para as categorias masculino e feminino. O padrão é \code{c("darkblue", "violetred")}.}

\item{nsize}{Tamanho do texto com o nº total de habitantes. O padrão é 3.5.}
}
\description{
Pirâmides populacionais com os arquivos de população
disponibilizados pelo DATASUS
}
\examples{
data("POPBR12")
str(POPBR12)
# "SEXO" é da classe 'integer'; 
# além disso, os primeiros 19 anos de idade estão em faixas anuais 
# e os demais (até 80) em faixas quinquenais.
\dontrun{
ggplot_pir(POPBR12, "FXETARIA", "SEXO", "POPULACAO")
}
# não dá o gráfico desejado

# Transformando o sexo em fator e reclassificando a faixa etária:
require(dplyr)
POPBR12 <- POPBR12 \%>\% 
  mutate(SEXO = factor(SEXO, labels = c("masc", "fem")),
         FXETAR5 = fxetar.det_pra_fxetar5(FXETARIA)) 
ggplot_pir(POPBR12, "FXETAR5", "SEXO", "POPULACAO")

# Cerro Largo
ggplot_pir(POPBR12[POPBR12$MUNIC_RES==430520, ], 
           "FXETAR5", "SEXO", "POPULACAO")


## Mortalidade por anos completos de vida

data("obitosRS2019")
ggplot_pir(obitosRS2019, "idade", "sexo", catsexo = 1:2)

}
\author{
Fúlvio B. Nedel
}
