% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histobox.R
\name{histobox}
\alias{histobox}
\title{Histograma com boxplot}
\usage{
histobox(x, limites = NULL, col.h = "cyan", col.bx = "skyblue1", ...)
}
\arguments{
\item{x}{uma variável numérica}

\item{limites}{limites do eixo x}

\item{col.h}{= cor do histograma}

\item{col.bx}{= cor do boxplot}

\item{...}{outros parâmetros de \code{\link{hist}} e \code{\link{boxplot}}}
}
\description{
Desenha um histograma com boxplot integrado ao mesmo gráfico
}
\examples{
histobox(rnorm(1000))
}
