% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histobox.R
\name{histobox}
\alias{histobox}
\title{Histograma com boxplot}
\usage{
histobox(
  x,
  limites = NULL,
  col.h = "cyan",
  col.bx = "skyblue1",
  maresq = 4.5,
  mardir = 2.1,
  ...
)
}
\arguments{
\item{x}{Uma variável numérica}

\item{limites}{Limites do eixo x}

\item{col.h}{Cor do histograma}

\item{col.bx}{Cor do boxplot}

\item{maresq}{Margem esquerda do gráfico}

\item{mardir}{Margem direita do gráfico}

\item{...}{Outros parâmetros de \code{\link{hist}} e \code{\link{boxplot}}}
}
\value{
O desenho do gráfico
}
\description{
Desenha um histograma com boxplot integrado ao mesmo gráfico
}
\examples{
varnum <- rnorm(1000)
histobox(varnum)
histobox(varnum, ylab = "Frequência")
histobox(varnum, maresq = 2.8)
# Se a margem esquerda for zero, o eixo y é removido:
histobox(varnum, col.h = "tomato", col.bx = "yellow", maresq = 0)
histobox(varnum, col.h = "tomato", col.bx = "yellow", mardir = 0)
histobox(varnum, col.h = "tomato", col.bx = "yellow", maresq = 0, mardir = 0)

}
