% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numescrito.R
\name{numescrito}
\alias{numescrito}
\title{Escreve números em texto, em português}
\usage{
numescrito(x, genero = "masc")
}
\arguments{
\item{x}{Número a ser convertido}

\item{genero}{Gênero do número. Se "fem" ou "f" o número será escrito no feminino.}
}
\description{
Dado um número entre 0 e 12, devolve seu valor em texto em português.
}
\details{
Para números negativos ou maiores de 12
}
\examples{
numescrito(0)
numescrito(1)
numescrito(1, genero = 'fem')
paste(numescrito(2, 'f'), "+", numescrito(3), "=", numescrito(2+3))
paste(numescrito(3), "-", numescrito(2, 'f'), "=", numescrito(3-2, 'f'))
# Limitações: 0<x<12
numescrito(-2)
numescrito(13)
numescrito(20)
paste(numescrito(6), "+", numescrito(7), "=", numescrito(6+7))
paste(numescrito(13), "-", numescrito(6), "=", numescrito(13-6))

}
