% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{obitosRS2019}
\alias{obitosRS2019}
\title{Registros de óbito}
\format{
Banco de dados com amostra aleatória de 10\% das DO de residentes do RS ocorridas em 2019, registrados no Sistema de Informação de Mortalidade (SIM). O banco foi previamente trabalhado para a decodificação da idade, de modo que a idade represente a idade em anos completos (0 para < 1 ano), e para a definição de missings e rótulos no campo sexo.
}
\source{
Sistema de Informações sobre Mortalidade (SIM)
}
\usage{
obitosRS2019
}
\description{
Declarações de óbito (DO) de residentes do RS, 2019.
}
\keyword{datasets}
