% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pir.R
\name{plot_pir}
\alias{plot_pir}
\title{Pirâmides populacionais com os arquivos de população
disponibilizados pelo DATASUS}
\usage{
plot_pir(
  pop,
  tabela = FALSE,
  ano = NULL,
  local = "população",
  title = NULL,
  fontsize = 1.1,
  nfxetar = 17,
  axes = FALSE,
  fonte = NULL,
  x.lim = NULL,
  colmasc = "mediumblue",
  colfem = "red2",
  colfxetar = "white",
  border = par("fg"),
  inside = T,
  drop.unused.levels = FALSE,
  ...
)
}
\arguments{
\item{pop}{Um \code{data frame} com a estrutura dos arquivos "POPBR??.DBF" disponibilizados pelo DATASUS, ou uma tabela com o sexo nas colunas (masc, fem) e a idade em 17 faixas etárias (0-4, ... 75-79, 80+) nas linhas}

\item{tabela}{Argumento lógico. Padrão é \code{FALSE}, deve ser mudado para \code{TRUE} quando a população é uma tabela como indicado acima}

\item{ano}{Indicar o período de referência. Apenas para quando a população é uma tabela}

\item{local}{Para o título: de onde é a população representada?}

\item{title}{Título do gáfico}

\item{fontsize}{Tamanho de fonte do título do gráfico}

\item{nfxetar}{No. de faixas etárias a serem representadas. Por enquanto única opção é 17 (quinquenais até 80 e +), tá na espera pra incluir outras opções}

\item{axes}{Não sei se vale a pena ir como opção ou já colocar direto o padrão FALSE. **Ver melhor**}

\item{fonte}{Texto para citar a fonte ao pé do gráfico}

\item{x.lim}{Limite do eixo x}

\item{colmasc}{Cor para o sexo masculino}

\item{colfem}{Cor para o sexo feminino}

\item{colfxetar}{Cor para os rótulos da faixa etária}

\item{border}{Cor da borda das barras}

\item{inside}{Ver a função \code{\link{barplot}}}

\item{drop.unused.levels}{Apagar níveis não usados nos fatores?}

\item{...}{argumentos de outras funções}
}
\description{
Pirâmides populacionais com os arquivos de população
disponibilizados pelo DATASUS
}
\examples{
\dontrun{
data("POPBR12")
plot_pir(POPBR12, local = 'Brasil')
plot_pir(POPBR12[substr(POPBR12$MUNIC_RES, 1,2)==42, ], local='Santa Catarina')
plot_pir(POPBR12[substr(POPBR12$MUNIC_RES, 1,2)==43, ], local='Rio Grande do Sul')
plot_pir(POPBR12[POPBR12$MUNIC_RES==431490, ], local='Porto Alegre, RS')
plot_pir(POPBR12[POPBR12$MUNIC_RES==430520, ], local='Cerro Largo, RS')
}

}
\author{
Fúlvio B. Nedel
}
\concept{csapAIH}
