% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZ.R
\name{plotZ}
\alias{plotZ}
\title{Gr치fico da probababilidade de pertencer a uma 치rea da curva Normal}
\usage{
plotZ(
  x = NULL,
  mu = 0,
  dp = 1,
  p = NULL,
  z = NULL,
  cor = 2,
  main = NULL,
  sub = NULL,
  sentido = "menor",
  cex.main = 2,
  cex.sub = 1.5,
  cex.axis = 1.3,
  ...
)
}
\description{
Gr치fico da probababilidade de pertencer a uma 치rea da curva Normal
}
\examples{
plotZ(p = .975)
plotZ(p = .025)
plotZ(z = 1.96)
plotZ(z = -1.96)
plotZ(p =  .975, sentido = "maior")
plotZ(p =  .025, sentido = "maior")
plotZ(z =  1.96, sentido = "maior")
plotZ(z = -1.96, sentido = "maior")

}
