% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.descreve.R
\name{print.descreve}
\alias{print.descreve}
\title{Imprime o resultado da funcao \code{\link{descreve}}}
\usage{
\method{print}{descreve}(x, dec = 2, ...)
}
\arguments{
\item{x}{Um output de \code{\link{descreve}}}

\item{dec}{Número de decimais}

\item{...}{**Acho** que permite coisas de \code{\link{print}}}
}
\description{
Imprime o resultado da funcao \code{\link{descreve}}
}
\concept{descreve}
