% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_csapAIH.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fxetar_quinq}
\alias{ufbr}
\alias{ler_popbr}
\title{Objects exported from other packages}
\seealso{
Other csapAIH: 
\code{\link{plot_pir}()}
}
\concept{DATASUS

Transforma um vetor com a idade em faixas etárias quinquenais. A primeira faixa pode ser fracionada em "< 1" e "1-4".}
\concept{csapAIH}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{csapAIH}{\code{\link[csapAIH]{fxetar_quinq}}, \code{\link[csapAIH]{ler_popbr}}, \code{\link[csapAIH]{ufbr}}}
}}

