% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacamostras.R
\name{sacamostras}
\alias{sacamostras}
\title{Toma múltiplas amostras aleatórias de um banco de dados.}
\usage{
sacamostras(
  data,
  size,
  prefix = NULL,
  suffixes,
  tipo = 2,
  seed = NULL,
  nomelongo = FALSE,
  csv = TRUE,
  rdata = TRUE
)
}
\arguments{
\item{data}{"Data frame" fonte das amostras.}

\item{size}{Tamanho das amostras (nº de registros).}

\item{prefix}{Texto para o início do nome dos "data frames" e bancos de dados.}

\item{suffixes}{Vetor numérico ou de caracteres, será usado como texto para o final do nome dos bancos de dados. Se for um vetor numérico e o argumento \code{seed} não for definido, será então usado para definir a "semente" das amostras (v. argumento \code{seed}).}

\item{tipo}{Formatação do arquivo csv; `tipo = 1` invoca a função \code{\link{write.csv}}, com campos separados por vírgula (`,`) e decimais separados por ponto (`.`), enquanto `tipo = 2` (padrão) invoca a função \code{\link{write.csv2}}, criando arquivos com campos separados por ponto-e-vírgula (`;`) e decimais separados por vírgula (`,`).}

\item{seed}{Vetor de números inteiros do tamanho do número de amostras desejado. É usado como "semente" para os números (pseudo)aleatórios que geram as amostras, permitindo assim sua reprodução. Se \code{NULL} (padrão), é tomado de \code{suffixes}, caso este seja um vetor de números inteiros. Argumento obrigatório quando \code{suffixes} não for numérico.}

\item{nomelongo}{Argumento lógico, FALSE por padrão. Se TRUE, acrescenta o número da semente ao nome do objeto e arquivo criados.}

\item{csv}{Argumento lógico, TRUE por padrão. Se FALSE, não cria o arquivo csv com a amostra .}

\item{rdata}{Argumento lógico, TRUE por padrão. Se FALSE, não cria o arquivo RData com a amostra .}
}
\description{
Toma múltiplas amostras aleatórias sem reposição de um banco de dados. Cada amostra é salva em um objeto da classe `data.frame` e exportada para um arquivo .csv, enquanto o conjunto dos objetos é salvo num arquivo de dados do R (.RData). Os arquivos são salvos no diretório de trabalho da sessão ativa. O nº de amostras a extrair é dado pelo comprimento de uma lista indicada no argumento `suffixes`
e cria novos "data frames" em um número definido de amostras aleatórias de tamanho n, criando com cada amostra um objeto da classe `data.frame`, que será salvo em um arquivo .csv e com os demais data frames, em um arquivo .RData. Criada com ajuda do chatGPT em 06/06/2023.
}
\examples{
# Perceba que a função não é enderaçada a nenhum objeto (como em `x <- sacamostras(...)`), 
# uma vez que ela já cria os bancos de dados como objetos no espaço de trabalho.
semente = 1:3
# Amostras de 100 registros:
sacamostras(data = RDRS2019, size = 100, prefix = "amostra_", suffixes = semente) 
# Amostras de 1\\% dos registros:
sacamostras(data = RDRS2019, size = .01, prefix = "amostra_", suffixes = semente) 
# Amostras de 1\\% dos registros, com outra semente:
sacamostras(data = RDRS2019, size = .01, prefix = "s", suffixes = semente, seed = 11:13)

# Amostras de 1\\% dos registros, com outros nomes, mas mesma "semente":
sacamostras(data = RDRS2019, size = .01, 
            prefix = "amostra", suffixes = paste0("0", 1:3), 
            seed = semente)
all.equal(amostra01, amostra_1) 

# Amostra de 1\\% dos registros, com mesmos nomes mas outra "semente":
sacamostras(data = RDRS2019, size = .01, prefix = "amostra", suffixes = semente, seed = 11:13)
all.equal(amostra1, s1) 

# A função retorna um aviso de erro se o argumento 'suffixes' não tiver 
# o mesmo comprimento do argumento 'seed':
\dontrun{
sacamostras(data = RDRS2019, size = .01, prefix = "amostra_", suffixes = "bis", seed = semente) 
}
# Amostra de 1\\% dos registros:
sacamostras(data = RDRS2019, size = .01, prefix = "amostra_", 
            suffixes = rep("bis", length(semente)), seed = semente) 
all.equal(amostra_bis, amostra_1) 
# Para repetir amostras sem sobreescrever os objetos e arquivos criados, 
# usa-se o argumento 'nomelongo = TRUE'.
sacamostras(data = RDRS2019, size = .01, prefix = "amostra_", 
            suffixes = rep("bis", length(semente)), seed = semente, nomelongo = TRUE) 
all.equal(amostra_bis_1, amostra_1) 
sacamostras(data = RDRS2019, size = .01, prefix = "amostra", suffixes = semente, seed = 11:13)
all.equal(amostra1, s1) 
rm(list = ls(pattern = "amostra"))
rm(list = ls(pattern = "s"))
unlink(c("amostra*", "s*")) # apaga os arquivos criados

}
