% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabuleiro2.R
\name{tabuleiro2}
\alias{tabuleiro2}
\title{Tabela de frequências com separadores latinos}
\usage{
tabuleiro2(varcat, digits = 1)
}
\arguments{
\item{varcat}{Uma variável categórica}

\item{digits}{No. de decimais}
}
\description{
Tabela de frequências com separadores latinos
}
\examples{
# criar uma variável politômica
x <- cut(rnorm(1000), 3) # sem missings
tabuleiro2(x)
x[1:100] <- NA # gerar missings
tabuleiro2(x)

\dontrun{
knitr::kable(tabuleiro2(x), align = rep("r", 4))
}

}
